requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiohttp==3.7.4.post0
    # via arsenic
apscheduler==3.6.3
    # via python-telegram-bot
arsenic==20.9
    # via -r requirements.in
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via
    #   aiohttp
    #   arsenic
certifi==2020.12.5
    # via python-telegram-bot
chardet==4.0.0
    # via aiohttp
idna==3.1
    # via yarl
loguru==0.5.3
    # via -r requirements.in
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
python-dotenv==0.17.1
    # via -r requirements.in
python-telegram-bot==13.5
    # via -r requirements.in
pytz==2021.1
    # via
    #   apscheduler
    #   python-telegram-bot
    #   tzlocal
six==1.16.0
    # via apscheduler
structlog==20.2.0
    # via arsenic
tornado==6.1
    # via python-telegram-bot
typing-extensions==3.10.0.0
    # via aiohttp
tzlocal==2.1
    # via apscheduler
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
